refactor(schema): extract CallBotRequest/Response models into separate schema file
